+++ app/components/primer/button/base.rb
@@ -46,7 +46,7 @@ module Primer
         @system_arguments[:tag] = fetch_or_fallback(TAG_OPTIONS, tag, DEFAULT_TAG)
 
         if @system_arguments[:tag] == :button
-          @system_arguments[:type] = type
+          @system_arguments[:type] = fetch_or_fallback(TYPE_OPTIONS, type, DEFAULT_TYPE)
         else
           @system_arguments[:role] = :button
         end