+++ stories/primer/button/marketing_stories.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+class Primer::Button::MarketingStories < ViewComponent::Storybook::Stories
+  layout "storybook_preview"
+
+  story(:button_marketing) do
+    controls do
+      select(:scheme, Primer::Button::Marketing::SCHEME_OPTIONS, :primary)
+      select(:variant, Primer::Button::Marketing::VARIANT_OPTIONS, :default)
+      select(:tag, Primer::Button::Marketing::TAG_OPTIONS, :button)
+      select(:type, Primer::Button::Marketing::TYPE_OPTIONS, :button)
+    end
+
+    content do
+      "Click me"
+    end
+  end
+end